+++ Automate/Framework/MachineGroup.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Microsoft.Xna.Framework;
@@ -30,8 +31,8 @@ namespace Pathoschild.Stardew.Automate.Framework
         /// <summary>The storage manager for the group.</summary>
         protected readonly StorageManager StorageManager;
 
-        /// <summary>The backing field for <see cref="Tiles"/>.</summary>
-        protected HashSet<Vector2> TilesImpl;
+        /// <summary>The tiles covered by this machine group.</summary>
+        private readonly HashSet<Vector2> Tiles;
 
 
         /*********
@@ -45,9 +46,6 @@ namespace Pathoschild.Stardew.Automate.Framework
 
         /// <inheritdoc />
         public IContainer[] Containers { get; protected set; }
-
-        /// <inheritdoc />
-        public IReadOnlySet<Vector2> Tiles => this.TilesImpl;
 
         /// <inheritdoc />
         [MemberNotNullWhen(false, nameof(IMachineGroup.LocationKey))]
@@ -71,10 +69,18 @@ namespace Pathoschild.Stardew.Automate.Framework
             this.LocationKey = locationKey;
             this.Machines = machines.ToArray();
             this.Containers = containers.ToArray();
-            this.TilesImpl = new HashSet<Vector2>(tiles);
+            this.Tiles = new HashSet<Vector2>(tiles);
 
             this.IsJunimoGroup = this.Containers.Any(p => p.IsJunimoChest);
             this.StorageManager = buildStorage(this.Containers);
+        }
+
+        /// <inheritdoc />
+        public virtual IReadOnlySet<Vector2> GetTiles(string locationKey)
+        {
+            return this.LocationKey == locationKey
+                ? this.Tiles
+                : ImmutableHashSet<Vector2>.Empty;
         }
 
         /// <inheritdoc />